+++ sshd-sftp/src/main/java/org/apache/sshd/sftp/client/fs/SftpPathIterator.java
@@ -19,30 +19,65 @@
 
 package org.apache.sshd.sftp.client.fs;
 
+import java.io.IOException;
+import java.nio.file.DirectoryIteratorException;
+import java.nio.file.DirectoryStream;
+import java.nio.file.DirectoryStream.Filter;
 import java.nio.file.Path;
 import java.util.Iterator;
 import java.util.NoSuchElementException;
+import java.util.Objects;
 
 import org.apache.sshd.sftp.client.SftpClient;
 
 /**
+ * Implements and {@link Iterator} of {@link SftpPath}-s returned by a {@link DirectoryStream#iterator()} method.
+ *
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
 public class SftpPathIterator implements Iterator<Path> {
-    private final SftpPath p;
-    private final Iterator<? extends SftpClient.DirEntry> it;
-    private boolean dotIgnored;
-    private boolean dotdotIgnored;
-    private SftpClient.DirEntry curEntry;
+    protected final Iterator<? extends SftpClient.DirEntry> it;
+    protected boolean dotIgnored;
+    protected boolean dotdotIgnored;
+    protected SftpPath curEntry;
+
+    private final SftpPath path;
+    private DirectoryStream.Filter<? super Path> filter;
 
     public SftpPathIterator(SftpPath path, Iterable<? extends SftpClient.DirEntry> iter) {
-        this(path, (iter == null) ? null : iter.iterator());
+        this(path, iter, null);
+    }
+
+    public SftpPathIterator(SftpPath path, Iterable<? extends SftpClient.DirEntry> iter,
+                            DirectoryStream.Filter<? super Path> filter) {
+        this(path, (iter == null) ? null : iter.iterator(), filter);
     }
 
     public SftpPathIterator(SftpPath path, Iterator<? extends SftpClient.DirEntry> iter) {
-        p = path;
+        this(path, iter, null);
+    }
+
+    public SftpPathIterator(SftpPath path, Iterator<? extends SftpClient.DirEntry> iter,
+                            DirectoryStream.Filter<? super Path> filter) {
+        this.path = Objects.requireNonNull(path, "No root path provided");
+        this.filter = filter;
+
         it = iter;
-        curEntry = nextEntry();
+        curEntry = nextEntry(path, filter);
+    }
+
+    /**
+     * @return The root {@link SftpPath} for this directory iterator
+     */
+    public final SftpPath getRootPath() {
+        return path;
+    }
+
+    /**
+     * @return The original filter - may be {@code null} to indicate no filter
+     */
+    public final Filter<? super Path> getFilter() {
+        return filter;
     }
 
     @Override
@@ -56,12 +91,12 @@ public class SftpPathIterator implements Iterator<Path> {
             throw new NoSuchElementException("No next entry");
         }
 
-        SftpClient.DirEntry entry = curEntry;
-        curEntry = nextEntry();
-        return p.resolve(entry.getFilename());
+        SftpPath returnValue = curEntry;
+        curEntry = nextEntry(getRootPath(), getFilter());
+        return returnValue;
     }
 
-    private SftpClient.DirEntry nextEntry() {
+    protected SftpPath nextEntry(SftpPath root, DirectoryStream.Filter<? super Path> selector) {
         while ((it != null) && it.hasNext()) {
             SftpClient.DirEntry entry = it.next();
             String name = entry.getFilename();
@@ -70,7 +105,14 @@ public class SftpPathIterator implements Iterator<Path> {
             } else if ("..".equals(name) && (!dotdotIgnored)) {
                 dotdotIgnored = true;
             } else {
-                return entry;
+                SftpPath candidate = root.resolve(entry.getFilename());
+                try {
+                    if ((selector == null) || selector.accept(candidate)) {
+                        return candidate;
+                    }
+                } catch (IOException e) {
+                    throw new DirectoryIteratorException(e);
+                }
             }
         }
 
@@ -79,6 +121,6 @@ public class SftpPathIterator implements Iterator<Path> {
 
     @Override
     public void remove() {
-        throw new UnsupportedOperationException("newDirectoryStream(" + p + ") Iterator#remove() N/A");
+        throw new UnsupportedOperationException("newDirectoryStream(" + getRootPath() + ") Iterator#remove() N/A");
     }
 }